Ajout optimisation sous contrainte par hyperplan

+++ simplexe.cpp
@@ -5,7 +5,7 @@ Simplexe::Simplexe(int dim, const std::vector<std::vector<double>>& points)
     : dimension(dim), noeuds(points)
 {
     // Un simplexe de dimension d possède d + 1 nœuds
-    if (noeuds.size() != dimension + 1)
+    if (noeuds.size() != static_cast<std::size_t>(dimension + 1))
     {
         throw std::invalid_argument(
             "Un simplexe de dimension d doit avoir d+1 noeuds."
@@ -15,7 +15,7 @@ Simplexe::Simplexe(int dim, const std::vector<std::vector<double>>& points)
     // Chaque nœud doit avoir d coordonnées
     for (const auto& noeud : noeuds)
     {
-        if (noeud.size() != dimension)
+        if (noeud.size() != static_cast<std::size_t>(dimension))
         {
             throw std::invalid_argument(
                 "La dimension d'un noeud est incorrecte."
@@ -46,7 +46,7 @@ void Simplexe::setNoeud(
     int i,
     const std::vector<double>& point)
 {
-    if (point.size() != dimension)
+    if (point.size() != static_cast<std::size_t>(dimension))
     {
         throw std::invalid_argument(
             "Le point n'a pas la bonne dimension."
